+++ Apps/CesiumViewer/CesiumViewer.js
@@ -35,6 +35,7 @@ define([
         'Scene/Material',
         'Scene/ExtentPrimitive',
         'Scene/Polygon',
+        'Scene/createTangentSpaceDebugPrimitive',
         'Widgets/checkForChromeFrame',
         'Widgets/Viewer/Viewer',
         'Widgets/Viewer/viewerDragDropMixin',
@@ -76,6 +77,7 @@ define([
         Material,
         ExtentPrimitive,
         Polygon,
+        createTangentSpaceDebugPrimitive,
         checkForChromeFrame,
         Viewer,
         viewerDragDropMixin,
@@ -211,7 +213,8 @@ define([
         });
         var geometry2 = new GeometryInstance({
             geometry : new EllipsoidGeometry({
-                vertexFormat : PerInstanceColorAppearance.VERTEX_FORMAT,
+                vertexFormat : VertexFormat.ALL,
+//                vertexFormat : PerInstanceColorAppearance.VERTEX_FORMAT,
                 ellipsoid : new Ellipsoid(500000.0, 500000.0, 1000000.0)
             }),
             modelMatrix : Matrix4.multiplyByTranslation(Transforms.eastNorthUpToFixedFrame(
@@ -251,6 +254,11 @@ define([
         });
         scene.getPrimitives().add(primitive);
 
+        scene.getPrimitives().add(createTangentSpaceDebugPrimitive({
+            geometry : geometry2.geometry,
+            length : 10000.0,
+            modelMatrix : geometry2.modelMatrix
+        }));
 
         var m = new Material({
             context : viewer.scene.getContext(),